requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
defusedxml==0.7.1
    # via jira
deprecated==1.2.18
    # via
    #   limits
    #   opentelemetry-api
distro==1.9.0
    # via openai
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-limiter
flask-limiter==3.5.1
    # via -r requirements.in
google-api-core==2.24.2
    # via
    #   google-api-python-client
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-api-python-client==2.128.0
    # via -r requirements.in
google-auth==2.40.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
google-cloud-appengine-logging==1.6.1
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
greenlet==3.2.2
    # via sqlalchemy
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jira==3.8.0
    # via -r requirements.in
limits==5.1.0
    # via flask-limiter
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
openai==1.30.2
    # via -r requirements.in
opentelemetry-api==1.32.1
    # via google-cloud-logging
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via
    #   gunicorn
    #   jira
    #   limits
pillow==11.2.1
    # via jira
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==5.29.4
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.5
    # via openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pyparsing==3.2.3
    # via httplib2
requests==2.32.3
    # via
    #   google-api-core
    #   jira
    #   requests-oauthlib
    #   requests-toolbelt
    #   tweepy
requests-oauthlib==1.3.1
    # via
    #   google-auth-oauthlib
    #   jira
    #   tweepy
requests-toolbelt==1.0.0
    # via jira
rich==13.9.4
    # via flask-limiter
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.29
    # via -r requirements.in
tqdm==4.67.1
    # via openai
tweepy==4.14.0
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   anyio
    #   flask-limiter
    #   jira
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.4.0
    # via requests
werkzeug==3.1.3
    # via flask
wrapt==1.17.2
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
